+++ src/index.js
@@ -30,7 +30,7 @@ export { date } from './date';
 export { dispatch } from './dispatch';
 export { either, either3, either4, either5, either6, either7, either8, either9 } from './either';
 export { fail } from './fail';
-export { mapping } from './mapping';
+export { mapping, dict } from './mapping';
 export { maybe } from './maybe';
 export { nullable } from './nullable';
 export { integer, number, positiveInteger, positiveNumber } from './number';